Treat conda installing seperate from plugin installs
