+++ conda/plugins/installers/__init__.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 """Register the built-in env_installer hook implementations."""
 
-from . import conda, pypi
+from . import pip, uv
 
 #: The list of env_installer plugins for easier registration with pluggy
-plugins = [conda, pypi]
+plugins = [pip, uv]
